requirements.txt
gradio>=4.0.0
requests
duckduckgo_search>=6.0.0
python-dotenv>=1.0.0
